♻️🏢 Extract action
