+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.stand-configuration/route-params.ts
@@ -1,0 +1,5 @@
+import { zu } from "@animeaux/zod-utils";
+
+export const RouteParamsSchema = zu.object({
+  id: zu.string().uuid(),
+});
