+++ apps/admin/src/routes/_layout.show.exhibitors.$id.edit.stand-configuration/route.tsx
@@ -1,31 +1,17 @@
 import { Action } from "#core/actions";
 import { ErrorPage, getErrorTitle } from "#core/data-display/error-page";
-import { db } from "#core/db.server";
 import { PageLayout } from "#core/layout/page";
-import { Routes } from "#core/navigation";
 import { getPageTitle } from "#core/page-title";
-import { assertCurrentUserHasGroups } from "#current-user/groups.server";
-import { safeParseRouteParam, zu } from "@animeaux/zod-utils";
-import type { SubmissionResult } from "@conform-to/react";
 import { getFormProps } from "@conform-to/react";
-import { parseWithZod } from "@conform-to/zod";
-import { UserGroup } from "@prisma/client";
-import type { ActionFunctionArgs } from "@remix-run/node";
-import { json } from "@remix-run/node";
 import type { MetaFunction } from "@remix-run/react";
-import type { MergeExclusive } from "type-fest";
-import { ActionSchema } from "./action";
 import { FieldsetConfiguration } from "./fieldset-configuration";
 import { FieldsetPriceDetails } from "./fieldset-price-details";
 import { FieldsetStatus } from "./fieldset-status";
 import { FormProvider, useFormRoot } from "./form";
 import type { loader } from "./loader.server.js";
 
+export { action } from "./action.server.js";
 export { loader } from "./loader.server.js";
-
-export const RouteParamsSchema = zu.object({
-  id: zu.string().uuid(),
-});
 
 export const meta: MetaFunction<typeof loader> = ({ data }) => {
   return [
@@ -38,53 +24,6 @@ export const meta: MetaFunction<typeof loader> = ({ data }) => {
     },
   ];
 };
-
-type ActionData = MergeExclusive<
-  { redirectTo: string },
-  { submissionResult: SubmissionResult<string[]> }
->;
-
-export async function action({ request, params }: ActionFunctionArgs) {
-  const currentUser = await db.currentUser.get(request, {
-    select: { groups: true },
-  });
-
-  assertCurrentUserHasGroups(currentUser, [
-    UserGroup.ADMIN,
-    UserGroup.SHOW_ORGANIZER,
-  ]);
-
-  const routeParams = safeParseRouteParam(RouteParamsSchema, params);
-
-  const formData = await request.formData();
-
-  const submission = parseWithZod(formData, { schema: ActionSchema });
-
-  if (submission.status !== "success") {
-    return json<ActionData>(
-      { submissionResult: submission.reply() },
-      { status: 400 },
-    );
-  }
-
-  await db.show.exhibitor.updateStand(routeParams.id, {
-    chairCount: submission.value.chairCount,
-    dividerCount: submission.value.dividerCount,
-    dividerTypeId: submission.value.dividerType ?? null,
-    hasElectricalConnection: submission.value.hasElectricalConnection,
-    hasTableCloths: submission.value.hasTableCloths,
-    installationDay: submission.value.installationDay,
-    peopleCount: submission.value.peopleCount,
-    sizeId: submission.value.sizeId,
-    standConfigurationStatus: submission.value.status,
-    standConfigurationStatusMessage: submission.value.statusMessage || null,
-    tableCount: submission.value.tableCount,
-  });
-
-  return json<ActionData>({
-    redirectTo: Routes.show.exhibitors.id(routeParams.id).toString(),
-  });
-}
 
 export function ErrorBoundary() {
   return (
